Filling edit course

+++ database/migrations/2019_04_28_205358_create_courses_table.php
@@ -13,17 +13,17 @@ class CreateCoursesTable extends Migration
      */
     public function up()
     {
-        Schema::create('courses', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('title')->nullable();
-            $table->string('image');
-            $table->text('short_description');
-            $table->text('description');
-            $table->tinyInteger('status')->default('0')->nullable();
-            $table->tinyInteger('kids')->default('0')->nullable();
-            $table->integer('priority')->default(10);
-            $table->timestamps();
-        });
+      Schema::create('courses', function (Blueprint $table) {
+        $table->increments('id');
+        $table->string('title')->nullable();
+        $table->string('image');
+        $table->text('short_description');
+        $table->text('description');
+        $table->tinyInteger('status')->default('0')->nullable();
+        $table->tinyInteger('kids')->default('0')->nullable();
+        $table->integer('priority')->default(10);
+        $table->timestamps();
+      });
     }
 
     /**
@@ -33,6 +33,6 @@ class CreateCoursesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('courses');
+      Schema::dropIfExists('courses');
     }
 }
